+++ frontend/srcs/src/app/app.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { CachedDataService } from './services/cached-data.service';
 
 @Component({
   selector: 'app-root',
@@ -7,12 +6,6 @@ import { CachedDataService } from './services/cached-data.service';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  constructor(
-    private readonly susbcribedDataService: CachedDataService
-    ) { }
   darkMode: boolean = true;
-
-
-
 
 }