+++ frontend/srcs/src/app/cache/channels.cache.ts
@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
-import { ChatService } from './chat.service';
-import { Channel, ChannelMessages, ChannelUsers, Chat, ChatMessages, User } from '../models';
+import { Channel, ChannelMessages, ChannelUsers } from '../models';
 import { Subject } from 'rxjs';
-import { ChannelService } from './channel.service';
-import { UserService } from './user.service';
+import { ChannelService } from '../services';
+
 interface MyCreatedAtObject {
   createdAt: string;
   // Otras propiedades que pueda tener su objeto
@@ -11,63 +10,14 @@ interface MyCreatedAtObject {
 @Injectable({
   providedIn: 'root'
 })
-export class CachedDataService {
-  private _conectedUsers = new Set<string>();
-
-  private _chatsAvailablesMap = new Map<string, Chat >();
-  private _chatsAvailables: Chat[] = [];
-  private _chatsMessages = new Map<string, ChatMessages[] >();
-
+export class ChannelsCache {
   private _joinedChannelsMap = new Map<string, Channel>();
   private _channelsMessages = new Map<string, ChannelMessages[]>();
   private _channelsUsers = new Map<string, ChannelUsers[]>();
 
-  private _cachedUsers = new Map<string, User>();
-
   constructor(
-    private readonly chatService: ChatService,
     private readonly channelService: ChannelService,
-    private readonly userService: UserService,
   ){
-    this.userService.usersConnected().subscribe(users => {
-        users.forEach(user =>{
-            this._conectedUsers.add(user);
-        });
-    });
-    this.userService.userDisconnects().subscribe(user => {
-        this._conectedUsers.delete(user);
-    });
-    this.userService.userConnects().subscribe(user => {
-        this._conectedUsers.add(user);
-    });
-
-    this.userService.usersToCache().subscribe(users => {
-      users.forEach(user =>{
-        user.avatar = `https://api.dicebear.com/7.x/bottts/svg?seed=${user.username}`;
-        if (this._cachedUsers.has(user.userId) === false) {
-          this._cachedUsers.set(user.userId, user);
-        }
-      });
-    });
-
-    this.chatService.chatsAvailables().subscribe(chats => {
-      chats.forEach(chat => {
-          this._chatsAvailablesMap.set(chat.chatId, chat);
-          this.chatService.chatLoaded(chat.chatId).subscribe(messages => {
-              this._setChatMessages(chat.chatId, messages);
-          });
-      });
-      this.updateChatsAvailablesSub();
-    });
-    this.chatService.newChatAvailable().subscribe(chat => {
-      if (!this._chatsAvailablesMap.has(chat.chatId)){
-          this._chatsAvailablesMap.set(chat.chatId, chat);
-          this.chatService.chatLoaded(chat.chatId).subscribe(messages => {
-              this._setChatMessages(chat.chatId, messages);
-          });
-          this.updateChatsAvailablesSub();
-      }
-    });
     this.channelService.joinedChannels().subscribe(channels => {
       channels.forEach(channel => {
           channel.avatar = `https://api.dicebear.com/7.x/bottts/svg?seed=${channel.channelName}`;  
@@ -80,21 +30,6 @@ export class CachedDataService {
           });
       });
     });
-  }
-
-  private _setChatMessages(chatId: string, messages: ChatMessages[]) {
-    let currentChatMsg: ChatMessages[] = this._chatsMessages.get(chatId) || [];
-    if (currentChatMsg?.length) {
-        messages.forEach(msg => {
-            if (currentChatMsg.find(x=>x.chatMessageId === msg.chatMessageId) === undefined) {
-                currentChatMsg.push(msg);
-            }
-        });
-        this._chatsMessages.set(chatId, currentChatMsg);
-    } else {
-        this._chatsMessages.set(chatId, messages);
-    }
-    this.updateChatMessagesSub(chatId, this._toDateMap(messages));
   }
 
   private _setChannelUsers(channelId: string, users: ChannelUsers[]) {
@@ -110,9 +45,9 @@ export class CachedDataService {
         this._channelsUsers.set(channelId, users);
     }
     this.updateChannelUsersSub(channelId, users);
-}
+  }
 
-private _setChannelMessages(channelId: string, messages: ChannelMessages[]) {
+  private _setChannelMessages(channelId: string, messages: ChannelMessages[]) {
     let currentChannelMsg: ChannelMessages[] = this._channelsMessages.get(channelId) || [];
     if (currentChannelMsg?.length) {
         messages.forEach(msg => {
@@ -125,51 +60,18 @@ private _setChannelMessages(channelId: string, messages: ChannelMessages[]) {
         this._channelsMessages.set(channelId, messages);
     }
     this.updateChannelMessagesSub(channelId, this._toDateMap(messages));
-}
+  }
 
   getJoinedChannels():Map< string, Channel >{
     return this._joinedChannelsMap
   }
 
-  getConnectedUsers(){
-    return this._conectedUsers;
-  }
-  isUserConnected(userId:string): boolean {
-    return this._conectedUsers.has(userId)
-  }
-
-  getUsername(userId: string): string {
-    return this._cachedUsers.get(userId)?.username || "noName"; 
-  }
-
-  getUser(userId: string): User | undefined {
-      return this._cachedUsers.get(userId) || undefined;
-  }
-
-  getUserImage(userId: string): string {
-    return this._cachedUsers.get(userId)?.avatar || "https://api.dicebear.com/avatar.svg"; 
-  }
-
-  getChatsAvailables(){
-    return this._chatsAvailables;
-  }
-  getChatsMessages(){
-    return this._chatsMessages;
-  }
   getChannelsMessages(){
     return this._channelsMessages;
   }
 
   getChannelsUsers(){
     return this._channelsUsers;
-  }
-
-  getChatMessages(chatId:string){
-    const messages = this._chatsMessages.get(chatId);
-    if ( messages !== undefined ) {
-      return this._toDateMap(messages);
-    }
-    return [];
   }
 
   getChannelMessages(channelId:string){
@@ -188,18 +90,6 @@ private _setChannelMessages(channelId: string, messages: ChannelMessages[]) {
     return [];
   }
 
-  private chatsAvailablesSub = new Subject<Chat[]>();
-  sendChatsAvailablesSub(data: Chat[]) {
-    this.chatsAvailablesSub.next(data);
-  }
-  getChatsAvailablesSub() {
-    return this.chatsAvailablesSub.asObservable();
-  }  
-  updateChatsAvailablesSub() {
-    this._chatsAvailables = Array.from(this._chatsAvailablesMap.values());
-    this.sendChatsAvailablesSub(this._chatsAvailables);
-  }
-
   private joinedChannelsSub = new Subject<Map< string, Channel >>();
   sendJoinedChannelsSub(data: Map< string, Channel >) {
     this.joinedChannelsSub.next(data);
@@ -209,17 +99,6 @@ private _setChannelMessages(channelId: string, messages: ChannelMessages[]) {
   }  
   updateJoinedChannelsSub() {
     this.sendJoinedChannelsSub(this._joinedChannelsMap);
-  }
-
-  private chatMessagesSub = new Subject< {chatId:string, chatMessages:Map<number, ChatMessages[]>} >();
-  sendChatMessagesSub(chatId:string, chatMessages:Map<number, ChatMessages[]>) {
-      this.chatMessagesSub.next({chatId, chatMessages});
-  }
-  getChatMessagesSub() {
-    return this.chatMessagesSub.asObservable();
-  }  
-  async updateChatMessagesSub(chatId:string, chatMessages: Map<number, ChatMessages[]>) {
-    this.sendChatMessagesSub(chatId, chatMessages);
   }
   
   private channelUsersSub = new Subject< {channelId:string, channelUsers:ChannelUsers[]} >();
@@ -243,7 +122,6 @@ private _setChannelMessages(channelId: string, messages: ChannelMessages[]) {
   async updateChannelMessagesSub(channelId:string, channelMessages: Map<number, ChannelMessages[]>) {
     this.sendChannelMessagesSub(channelId, channelMessages);
   }
-
 
   private _toDateMap<T extends MyCreatedAtObject>(items: T[]): Map<number,T[]> {
     let groupedItems: Map<number,T[]> = new Map<number,T[]>();