+++ frontend/srcs/src/app/cache/index.ts
@@ -1,0 +1,3 @@
+export * from './channels.cache'
+export * from './chats.cache'
+export * from './users.cache'