+++ frontend/srcs/src/app/components/chat/chat-info/chat-info.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+import { User } from 'src/app/models';
+
+
+@Component({
+  selector: 'app-chat-info',
+  templateUrl: './chat-info.component.html',
+  styleUrls: ['./chat-info.component.scss']
+})
+export class ChatInfoComponent {
+  @Input() user!: User;
+}