+++ frontend/srcs/src/app/components/chat/index.ts
@@ -1,0 +1,4 @@
+export * from './chat-info/chat-info.component'
+export * from './chat-sidebar/chat-sidebar.component'
+export * from './chat-window/chat-window.component'
+export * from './chat-window-message/chat-window-message.component'