+++ frontend/srcs/src/app/components/index.ts
@@ -1,0 +1,3 @@
+export * from './avatar/avatar.component'
+export * from './channel'
+export * from './chat'