+++ frontend/srcs/src/app/modules/index.ts
@@ -1,0 +1,7 @@
+export * from './chat/chat.component'
+export * from './game/game.component'
+export * from './home/home.component'
+export * from './profile/profile.component'
+export * from './history/history.component'
+export * from './shared'
+