+++ frontend/srcs/src/app/modules/shared/index.ts
@@ -1,4 +1,5 @@
 export * from './base/base.component';
+export * from './navbar/navbar.component'
 // export * from './footer/footer.module';
 // export * from './header/header.module';
 // export * from './alert/alert.module';