+++ frontend/srcs/src/app/services/index.ts
@@ -2,5 +2,4 @@ export * from './api.service';
 export * from './auth.service';
 export * from './chat.service';
 export * from './channel.service';
-export * from './cached-data.service';
 export * from './user.service';