+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-10-25 - emadriga
+### Changed
+- Reorganize front modules, components & services 
+
 ## 2023-10-24 - emadriga
 ### Changed
 - Decoupling user data from chat & channel backs